teste-tcc/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
colorama==0.4.6
fastapi==0.109.2
h11==0.14.0
idna==3.6
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
sniffio==1.3.0
starlette==0.36.3
tinydb==4.8.0
typing_extensions==4.9.0
uvicorn==0.27.0.post1